added reset_clock + clock changes

+++ src/ca/types.h
@@ -11,8 +11,8 @@ namespace ca
 	typedef int i32;
 	typedef unsigned int u32;
 
-	typedef long i64;
-	typedef unsigned long u64;
+	typedef long long i64;
+	typedef unsigned long long u64;
 
 	typedef float f32;
 	typedef double f64;
